+++ src/Local.h
@@ -12,24 +12,24 @@ using namespace std;
 
 class Local{
 private:
-	int id;
+	long long id;
 	pair<float, float> coordinates;		//latitude, longitude
 	int height;
-	map<int, string> roads;		//idRoad, roadName
+	map<long long, string> roads;		//idRoad, roadName
 	SharingPoint* sharingPoint = NULL;
 public:
-	Local(int id, pair<float, float> coord, int height);
-	int getId() const;
+	Local(long long id, pair<float, float> coord, int height);
+	long long getId() const;
 	pair<float, float> getCoordinates() const;
 	int getHeight() const;
 	float getDistance(const Local nd) const;
 	SharingPoint* getSharingPoint();
-	map<int, string> getRoads();
+	map<long long, string> getRoads();
 	bool operator ==(const Local & nd);
 	void print();
 	void setSharingPoint(SharingPoint* sp);
-	void addRoad(pair<int, string> road);
-	bool setRoadName(int roadId, string roadName);
+	void addRoad(pair<long long, string> road);
+	bool setRoadName(long long roadId, string roadName);
 };
